refactor for neg-root sync
