--- test/did_dht_test.go
@@ -1,1 +1,0 @@
-package test
